began debugging ising3d, and also catch up to 3Dxy with save etc

+++ modular/source/Ising3D/Ising3Dwolff.h
@@ -5,7 +5,7 @@
 #include "../randStruct.h"
 #include "../clusterStruct.h"
 
-long int clusterIsing3D(LatticeIsing3D& lat);
+long double clusterIsing3D(LatticeIsing3D& lat);
